+++ src/renderer/components/states/announcement/AnnouncementContent.tsx
@@ -27,14 +27,13 @@
 import { useState, useEffect, useCallback, useRef, type ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import useIslandStore from '../../../store/slices';
-import { QuestionnaireBanner } from '../../components/DynamicIslandQuestionnaireBanner';
+import { QuestionnaireBanner, useAnnouncementQuestionnaire } from '../../components/DynamicIslandQuestionnaireBanner';
 import { AnnouncementHeader } from './components/AnnouncementHeader';
 import { AnnouncementBody } from './components/AnnouncementBody';
 import { ANNOUNCEMENT_DEFAULTS, ANNOUNCEMENT_KEYS } from './config/announcementDefaults';
 import { useAdSlides } from './hooks/useAdSlides';
 import { AD_SLIDE_INTERVAL_MS } from './config/adSlidesDefaults';
 import { useAnnouncementData } from './hooks/useAnnouncementData';
-import { useAnnouncementQuestionnaire } from './hooks/useAnnouncementQuestionnaire';
 import { ProcessIndicator } from '../../components/DynamicIslandProcessIndicator';
 import { SvgIcon } from '../../../utils/SvgIcon';
 import '../../../styles/announcement/announcement.css';
