+++ src/renderer/components/states/maxExpand/components/setting/components/index/IndexSettingsSection.tsx
@@ -27,8 +27,7 @@
 import { useState, useMemo, type MutableRefObject, type ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import useIslandStore from '../../../../../../../store/slices';
-import { useAnnouncementQuestionnaire } from '../../../../../../states/announcement/hooks/useAnnouncementQuestionnaire';
-import { QuestionnaireBanner } from '../../../../../../../components/components/DynamicIslandQuestionnaireBanner';
+import { QuestionnaireBanner, useAnnouncementQuestionnaire } from '../../../../../../../components/components/DynamicIslandQuestionnaireBanner';
 import { SEARCHABLE_SETTINGS, SETTINGS_TAB_ICONS, type SearchableSettingItem, type AppSettingsPageKey, type MusicSettingsPageKey, type AiSettingsPageKey, type NetworkSettingsPageKey, type SettingsSidebarTabKey } from '../../utils/settingsConfig';
 import '../../../../../../../styles/announcement/announcement.css';
